requirements.txt
fastapi>=0.111.0
starlette>=0.37.2
uvicorn>=0.30.1
pydantic>=2.8.0
fastapi-restful>=0.6.0
typing-inspect>=0.9.0
SQLAlchemy>=2.0.31
asyncpg>=0.29.0
bcrypt>=4.1.3
PyYAML>=6.0.1
types-PyYAML>=6.0.12.20240311